requirements.txt
streamlit>=1.32.0
PyPDF2>=3.0.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
numpy>=1.24.0
groq>=0.5.0
google-generativeai>=0.5.0
